+++ modules/AWSPowerShell/Cmdlets/AmazonECS/Basic/Get-ECSContainerInstanceDetail-Cmdlet.cs
@@ -54,7 +54,7 @@ namespace Amazon.PowerShell.Cmdlets.ECS
         #region Parameter ContainerInstance
         /// <summary>
         /// <para>
-        /// <para>A list of container instance IDs or full ARN entries.</para>
+        /// <para>A list of up to 100 container instance IDs or full Amazon Resource Name (ARN) entries.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]