+++ modules/AWSPowerShell/Cmdlets/AmazonSimpleEmailService/Basic/Confirm-SESDomainIdentity-Cmdlet.cs
@@ -28,8 +28,9 @@ using Amazon.SimpleEmail.Model;
 namespace Amazon.PowerShell.Cmdlets.SES
 {
     /// <summary>
-    /// Adds a domain to the list of identities for your Amazon SES account and attempts to
-    /// verify it. For more information about verifying domains, see <a href="http://docs.aws.amazon.com/ses/latest/DeveloperGuide/verify-addresses-and-domains.html">Verifying
+    /// Adds a domain to the list of identities for your Amazon SES account in the current
+    /// AWS Region and attempts to verify it. For more information about verifying domains,
+    /// see <a href="http://docs.aws.amazon.com/ses/latest/DeveloperGuide/verify-addresses-and-domains.html">Verifying
     /// Email Addresses and Domains</a> in the <i>Amazon SES Developer Guide.</i><para>
     /// You can execute this operation no more than once per second.
     /// </para>