+++ modules/AWSPowerShell/Cmdlets/AmazonSimpleEmailService/Basic/Get-SESConfigurationSetList-Cmdlet.cs
@@ -28,8 +28,8 @@ using Amazon.SimpleEmail.Model;
 namespace Amazon.PowerShell.Cmdlets.SES
 {
     /// <summary>
-    /// Provides a list of the configuration sets associated with your Amazon SES account.
-    /// For information about using configuration sets, see <a href="http://docs.aws.amazon.com/ses/latest/DeveloperGuide/monitor-sending-activity.html">Monitoring
+    /// Provides a list of the configuration sets associated with your Amazon SES account
+    /// in the current AWS Region. For information about using configuration sets, see <a href="http://docs.aws.amazon.com/ses/latest/DeveloperGuide/monitor-sending-activity.html">Monitoring
     /// Your Amazon SES Sending Activity</a> in the <i>Amazon SES Developer Guide.</i><para>
     /// You can execute this operation no more than once per second. This operation will return
     /// up to 1,000 configuration sets each time it is run. If your Amazon SES account has