+++ modules/AWSPowerShell/Cmdlets/AmazonSimpleEmailService/Basic/Get-SESCustomVerificationEmailTemplateList-Cmdlet.cs
@@ -28,11 +28,12 @@ using Amazon.SimpleEmail.Model;
 namespace Amazon.PowerShell.Cmdlets.SES
 {
     /// <summary>
-    /// Lists the existing custom verification email templates for your account.
+    /// Lists the existing custom verification email templates for your account in the current
+    /// AWS Region.
     /// 
     ///  
     /// <para>
-    /// For more information about custom verification email templates, see <a href="https://docs.aws.amazon.com/ses/latest/DeveloperGuide/custom-verification-emails.html">Using
+    /// For more information about custom verification email templates, see <a href="http://docs.aws.amazon.com/ses/latest/DeveloperGuide/custom-verification-emails.html">Using
     /// Custom Verification Email Templates</a> in the <i>Amazon SES Developer Guide</i>.
     /// </para><para>
     /// You can execute this operation no more than once per second.