+++ modules/AWSPowerShell/Cmdlets/AmazonSimpleEmailService/Basic/Get-SESIdentity-Cmdlet.cs
@@ -29,7 +29,7 @@ namespace Amazon.PowerShell.Cmdlets.SES
 {
     /// <summary>
     /// Returns a list containing all of the identities (email addresses and domains) for
-    /// your AWS account, regardless of verification status.
+    /// your AWS account in the current AWS Region, regardless of verification status.
     /// 
     ///  
     /// <para>