+++ modules/AWSPowerShell/Cmdlets/AmazonSimpleEmailService/Basic/Get-SESReceiptRuleSetList-Cmdlet.cs
@@ -28,10 +28,10 @@ using Amazon.SimpleEmail.Model;
 namespace Amazon.PowerShell.Cmdlets.SES
 {
     /// <summary>
-    /// Lists the receipt rule sets that exist under your AWS account. If there are additional
-    /// receipt rule sets to be retrieved, you will receive a <code>NextToken</code> that
-    /// you can provide to the next call to <code>ListReceiptRuleSets</code> to retrieve the
-    /// additional entries.
+    /// Lists the receipt rule sets that exist under your AWS account in the current AWS Region.
+    /// If there are additional receipt rule sets to be retrieved, you will receive a <code>NextToken</code>
+    /// that you can provide to the next call to <code>ListReceiptRuleSets</code> to retrieve
+    /// the additional entries.
     /// 
     ///  
     /// <para>