+++ modules/AWSPowerShell/Cmdlets/AmazonSimpleEmailService/Basic/Get-SESSendStatistic-Cmdlet.cs
@@ -28,7 +28,7 @@ using Amazon.SimpleEmail.Model;
 namespace Amazon.PowerShell.Cmdlets.SES
 {
     /// <summary>
-    /// Provides sending statistics for the Amazon SES account. The result is a list of data
+    /// Provides sending statistics for the current AWS Region. The result is a list of data
     /// points, representing the last two weeks of sending activity. Each data point in the
     /// list contains statistics for a 15-minute period of time.
     ///