+++ modules/AWSPowerShell/Cmdlets/AmazonSimpleEmailService/Basic/Remove-SESConfigurationSetTrackingOption-Cmdlet.cs
@@ -35,8 +35,7 @@ namespace Amazon.PowerShell.Cmdlets.SES
     /// <para>
     /// By default, images and links used for tracking open and click events are hosted on
     /// domains operated by Amazon SES. You can configure a subdomain of your own to handle
-    /// these events. For information about using configuration sets, see <a href="http://docs.aws.amazon.com/ses/latest/DeveloperGuide/configure-custom-open-click-domains.html">Configuring
-    /// Custom Domains to Handle Open and Click Tracking</a> in the <a href="http://docs.aws.amazon.com/ses/latest/DeveloperGuide/Welcome.html">Amazon
+    /// these events. For information about using custom domains, see the <a href="http://docs.aws.amazon.com/ses/latest/DeveloperGuide/configure-custom-open-click-domains.html">Amazon
     /// SES Developer Guide</a>.
     /// </para><note><para>
     /// Deleting this kind of association will result in emails sent using the specified configuration