+++ modules/AWSPowerShell/Cmdlets/AmazonSimpleEmailService/Basic/Update-SESAccountSendingEnabled-Cmdlet.cs
@@ -28,10 +28,10 @@ using Amazon.SimpleEmail.Model;
 namespace Amazon.PowerShell.Cmdlets.SES
 {
     /// <summary>
-    /// Enables or disables email sending across your entire Amazon SES account. You can use
-    /// this operation in conjunction with Amazon CloudWatch alarms to temporarily pause email
-    /// sending across your Amazon SES account when reputation metrics (such as your bounce
-    /// on complaint rate) reach certain thresholds.
+    /// Enables or disables email sending across your entire Amazon SES account in the current
+    /// AWS Region. You can use this operation in conjunction with Amazon CloudWatch alarms
+    /// to temporarily pause email sending across your Amazon SES account in a given AWS Region
+    /// when reputation metrics (such as your bounce or complaint rates) reach certain thresholds.
     /// 
     ///  
     /// <para>
@@ -51,7 +51,8 @@ namespace Amazon.PowerShell.Cmdlets.SES
         #region Parameter Enabled
         /// <summary>
         /// <para>
-        /// <para>Describes whether email sending is enabled or disabled for your Amazon SES account.</para>
+        /// <para>Describes whether email sending is enabled or disabled for your Amazon SES account
+        /// in the current AWS Region.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(Position = 0, ValueFromPipeline = true)]