+++ modules/AWSPowerShell/Cmdlets/AmazonSimpleEmailService/Basic/Update-SESConfigurationSetSendingEnabled-Cmdlet.cs
@@ -29,9 +29,10 @@ namespace Amazon.PowerShell.Cmdlets.SES
 {
     /// <summary>
     /// Enables or disables email sending for messages sent using a specific configuration
-    /// set. You can use this operation in conjunction with Amazon CloudWatch alarms to temporarily
-    /// pause email sending for a configuration set when the reputation metrics for that configuration
-    /// set (such as your bounce on complaint rate) reach certain thresholds.
+    /// set in a given AWS Region. You can use this operation in conjunction with Amazon CloudWatch
+    /// alarms to temporarily pause email sending for a configuration set when the reputation
+    /// metrics for that configuration set (such as your bounce on complaint rate) exceed
+    /// certain thresholds.
     /// 
     ///  
     /// <para>